frontend/requirements.txt
streamlit
requests
python-dotenv
flake8
pytest
